# openapi/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "vicinopoli",
    "version": "0.1.0"
  },
  "paths": {
    "/api/admin/posts": {
      "get": {
        "tags": [
          "admin"
        ],
        "summary": "Get Admin Posts",
        "description": "Return a page of all posts, newest first, with report counts.\n\nKeyset pagination mirrors the public feed: ``next_cursor`` encodes the last\npost's ``(created_at, id)`` and the next request resumes strictly after it.",
        "operationId": "get_admin_posts_api_admin_posts_get",
        "parameters": [
          {
            "name": "limit",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "maximum": 100,
              "minimum": 1,
              "default": 20,
              "title": "Limit"
            }
          },
          {
            "name": "cursor",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Cursor"
            }
          },
          {
            "name": "x-admin-token",
            "in": "header",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "X-Admin-Token"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AdminFeedResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/admin/push/inbox": {
      "get": {
        "tags": [
          "admin"
        ],
        "summary": "Push Inbox Read",
        "description": "Return every mock push delivery (optionally clearing on read).",
        "operationId": "push_inbox_read_api_admin_push_inbox_get",
        "parameters": [
          {
            "name": "clear",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "default": 0,
              "title": "Clear"
            }
          },
          {
            "name": "x-admin-token",
            "in": "header",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "X-Admin-Token"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "type": "object",
                    "additionalProperties": true
                  },
                  "title": "Response Push Inbox Read Api Admin Push Inbox Get"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      },
      "post": {
        "tags": [
          "admin"
        ],
        "summary": "Push Inbox Write",
        "description": "Record a delivery, keyed by the ``to`` query parameter (if any).",
        "operationId": "push_inbox_write_api_admin_push_inbox_post",
        "parameters": [
          {
            "name": "x-admin-token",
            "in": "header",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "X-Admin-Token"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "additionalProperties": true,
                "title": "Payload"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "additionalProperties": true,
                  "title": "Response Push Inbox Write Api Admin Push Inbox Post"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
    "/api/posts": {
      "post": {
        "tags": [
          "posts"
        ],
        "summary": "Create Post",
        "operationId": "create_post_api_posts_post",
        "parameters": [
          {
            "name": "device_id",
            "in": "cookie",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Device Id"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/PostCreate"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PostResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/feed": {
      "get": {
        "tags": [
          "posts"
        ],
        "summary": "Get Feed",
        "operationId": "get_feed_api_feed_get",
        "parameters": [
          {
            "name": "address",
            "in": "query",
            "required": true,
            "schema": {
              "type": "string",
              "minLength": 1,
              "maxLength": 512,
              "title": "Address"
            }
          },
          {
            "name": "target_count",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "maximum": 50,
              "minimum": 1,
              "default": 10,
              "title": "Target Count"
            }
          },
          {
            "name": "cursor",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Cursor"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/FeedResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
      "AdminFeedResponse": {
        "properties": {
          "posts": {
            "items": {
              "$ref": "#/components/schemas/AdminPost"
            },
            "type": "array",
            "title": "Posts"
          },
          "next_cursor": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Next Cursor"
          }
        },
        "type": "object",
        "required": [
          "posts"
        ],
        "title": "AdminFeedResponse"
      },
      "AdminPost": {
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid",
            "title": "Id"
          },
          "body": {
            "type": "string",
            "title": "Body"
          },
          "voice": {
            "$ref": "#/components/schemas/PostVoice"
          },
          "status": {
            "$ref": "#/components/schemas/PostStatus"
          },
          "display_address": {
            "type": "string",
            "title": "Display Address"
          },
          "geohash": {
            "type": "string",
            "title": "Geohash"
          },
          "created_at": {
            "type": "string",
            "format": "date-time",
            "title": "Created At"
          },
          "pseudonym": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Pseudonym"
          },
          "new_neighbour": {
            "type": "boolean",
            "title": "New Neighbour",
            "default": true
          },
          "report_count": {
            "type": "integer",
            "title": "Report Count",
            "default": 0
          },
          "device_id": {
            "anyOf": [
              {
                "type": "string",
                "format": "uuid"
              },
              {
                "type": "null"
              }
            ],
            "title": "Device Id"
          },
          "media": {
            "items": {
              "$ref": "#/components/schemas/MediaInfo"
            },
            "type": "array",
            "title": "Media"
          }
        },
        "type": "object",
        "required": [
          "id",
          "body",
          "voice",
          "status",
          "display_address",
          "geohash",
          "created_at"
        ],
        "title": "AdminPost"
      },
...
      "FeedItem": {
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid",
            "title": "Id"
          },
          "body": {
            "type": "string",
            "title": "Body"
          },
          "scope": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/PostScope"
              },
              {
                "type": "null"
              }
            ]
          },
          "voice": {
            "$ref": "#/components/schemas/PostVoice"
          },
          "display_address": {
            "type": "string",
            "title": "Display Address"
          },
          "geohash": {
            "type": "string",
            "title": "Geohash"
          },
          "distance_m": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Distance M"
          },
          "created_at": {
            "type": "string",
            "format": "date-time",
            "title": "Created At"
          },
          "pseudonym": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Pseudonym"
          },
          "new_neighbour": {
            "type": "boolean",
            "title": "New Neighbour",
            "default": true
          },
          "media": {
            "items": {
              "$ref": "#/components/schemas/MediaInfo"
            },
            "type": "array",
            "title": "Media"
          }
        },
        "type": "object",
        "required": [
          "id",
          "body",
          "voice",
          "display_address",
          "geohash",
          "created_at"
        ],
        "title": "FeedItem"
      },
      "FeedResponse": {
        "properties": {
          "posts": {
            "items": {
              "$ref": "#/components/schemas/FeedItem"
            },
            "type": "array",
            "title": "Posts"
          },
          "effective_radius_m": {
            "type": "integer",
            "title": "Effective Radius M"
          },
          "target_count": {
            "type": "integer",
            "title": "Target Count"
          },
          "next_cursor": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Next Cursor"
          }
        },
        "type": "object",
        "required": [
          "posts",
          "effective_radius_m",
          "target_count"
        ],
        "title": "FeedResponse"
      },
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
...
      "LocationInfo": {
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid",
            "title": "Id"
          },
          "display_address": {
            "type": "string",
            "title": "Display Address"
          },
          "geohash": {
            "type": "string",
            "title": "Geohash"
          }
        },
        "type": "object",
        "required": [
          "id",
          "display_address",
          "geohash"
        ],
        "title": "LocationInfo"
      },
      "MediaInfo": {
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid",
            "title": "Id"
          },
          "kind": {
            "type": "string",
            "title": "Kind"
          },
          "url": {
            "type": "string",
            "title": "Url"
          },
          "duration_s": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Duration S"
          }
        },
        "type": "object",
        "required": [
          "id",
          "kind",
          "url"
        ],
        "title": "MediaInfo"
      },
...
      "PostCreate": {
        "properties": {
          "address": {
            "type": "string",
            "maxLength": 512,
            "minLength": 1,
            "title": "Address"
          },
          "body": {
            "type": "string",
            "maxLength": 5000,
            "title": "Body",
            "default": ""
          },
          "voice": {
            "$ref": "#/components/schemas/PostVoice",
            "default": "city"
          },
          "scope": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/PostScope"
              },
              {
                "type": "null"
              }
            ]
          },
          "media_ids": {
            "items": {
              "type": "string",
              "format": "uuid"
            },
            "type": "array",
            "maxItems": 9,
            "title": "Media Ids"
          }
        },
        "type": "object",
        "required": [
          "address"
        ],
        "title": "PostCreate"
      },
      "PostResponse": {
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid",
            "title": "Id"
          },
          "body": {
            "type": "string",
            "title": "Body"
          },
          "scope": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/PostScope"
              },
              {
                "type": "null"
              }
            ]
          },
          "voice": {
            "$ref": "#/components/schemas/PostVoice"
          },
          "location": {
            "$ref": "#/components/schemas/LocationInfo"
          },
          "distance_m": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Distance M"
          },
          "created_at": {
            "type": "string",
            "format": "date-time",
            "title": "Created At"
          },
          "pseudonym": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Pseudonym"
          },
          "new_neighbour": {
            "type": "boolean",
            "title": "New Neighbour",
            "default": true
          },
          "daily_post_limit": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Daily Post Limit"
          },
          "posts_left_today": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Posts Left Today"
          },
          "media": {
            "items": {
              "$ref": "#/components/schemas/MediaInfo"
            },
            "type": "array",
            "title": "Media"
          }
        },
        "type": "object",
        "required": [
          "id",
          "body",
          "voice",
          "location",
          "created_at"
        ],
        "title": "PostResponse"
      },
      "PostScope": {
        "type": "string",
        "enum": [
          "street",
          "500m",
          "1km",
          "5km"
        ],
        "title": "PostScope"
      },
      "PostStatus": {
        "type": "string",
        "enum": [
          "active",
          "auto_hidden",
          "hidden"
        ],
        "title": "PostStatus"
      },
      "PostVoice": {
        "type": "string",
        "enum": [
          "street",
          "some",
          "area",
          "city"
        ],
        "title": "PostVoice",
        "description": "The author's voice intent (fuzzy, user-facing): who should read this.\n\nThe voice is stored on the post as-is; the feed converts it to a fixed\n``reach`` (a distance) via ``VOICE_TO_REACH_M`` (glossary: reach, ADR\n0024): ``street`` -> 5m, ``some`` -> 500m, ``area`` -> 3km, ``city`` ->\n50km. ``street`` is a 5m radius, not a normalized address key match."
      },
...
      "ValidationError": {
        "properties": {
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "type": "array",
            "title": "Location"
          },
          "msg": {
            "type": "string",
            "title": "Message"
          },
          "type": {
            "type": "string",
            "title": "Error Type"
          },
          "input": {
            "title": "Input"
          },
          "ctx": {
            "type": "object",
            "title": "Context"
          }
        },
        "type": "object",
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError"
      }
    }
  }
}